% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_factor_type}
\alias{detect_factor_type}
\title{Detect whether a variable in a model is continuous or categorical}
\usage{
detect_factor_type(model, varname)
}
\arguments{
\item{model}{A fitted model object.}

\item{varname}{Character string naming the variable.}
}
\value{
Character: "continuous" or "categorical".
}
\description{
Detect whether a variable in a model is continuous or categorical
}
\keyword{internal}
