% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_residuals.R
\name{.extract_caret_cv}
\alias{.extract_caret_cv}
\title{Extract CV residuals from a caret train object}
\usage{
.extract_caret_cv(model, data, response)
}
\description{
Extract CV residuals from a caret train object
}
\keyword{internal}
