% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_dominance.R
\name{.point_in_polygon}
\alias{.point_in_polygon}
\title{Point-in-polygon test using ray casting}
\usage{
.point_in_polygon(px, py, vx, vy)
}
\description{
Point-in-polygon test using ray casting
}
\keyword{internal}
