% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ixsurface-package.R
\docType{package}
\name{ixsurface-package}
\alias{ixsurface}
\alias{ixsurface-package}
\title{ixsurface: Interactive 3D Surface Plots for Multi-Factor Interaction Visualization}
\description{
Visualize interactions between multiple experimental factors using interactive 3D surface plots powered by 'plotly'. Instead of examining combinatorial pairwise interaction plots, map factor combinations to response surfaces and use surface crossings as geometric indicators of interaction effects. Supports continuous, categorical, and mixed factor designs with automatic binning for continuous conditioning variables.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cjbrant/ixsurface}
  \item Report bugs at \url{https://github.com/cjbrant/ixsurface/issues}
}

}
\author{
\strong{Maintainer}: Chris Brantner \email{cjbrantn@gmail.com}

Authors:
\itemize{
  \item Chris Brantner \email{cjbrantn@gmail.com}
}

}
\keyword{internal}
