% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lighten_color}
\alias{lighten_color}
\title{Lighten a hex color}
\usage{
lighten_color(hex_color, amount = 0.3)
}
\arguments{
\item{hex_color}{Character hex color.}

\item{amount}{Numeric in \code{[0, 1]}. How much to lighten.}
}
\value{
Character hex color.
}
\description{
Lighten a hex color
}
\keyword{internal}
