% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_by_label}
\alias{make_by_label}
\title{Generate a label for a facet_by combination}
\usage{
make_by_label(row, by_vars, binned_by = list())
}
\arguments{
\item{row}{A single-row data.frame.}

\item{by_vars}{Character vector of by-variable names.}

\item{binned_by}{Named list of bin info for continuous by-variables.}
}
\value{
A character label.
}
\description{
Generate a label for a facet_by combination
}
\keyword{internal}
