% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{surface_palette}
\alias{surface_palette}
\title{Get a qualitative color palette for surfaces}
\usage{
surface_palette(n)
}
\arguments{
\item{n}{Number of colors needed.}
}
\value{
Character vector of hex colors.
}
\description{
Get a qualitative color palette for surfaces
}
\keyword{internal}
